+++ AMButtonBar/NSShadow_AMAdditions.h
@@ -4,14 +4,12 @@
 //
 //  Created by Andreas on 10.02.07.
 //  Copyright 2007 Andreas Mayer. All rights reserved.
+//  Copyright 2010 Hardcoded Software (http://www.hardcoded.net)
 //
 
 #import <Cocoa/Cocoa.h>
 
 
 @interface NSShadow (AMAdditions)
-
-+ (NSShadow *)shadowWithColor:(NSColor *)color blurRadius:(float)radius offset:(NSSize)offset;
-
-
++ (NSShadow *)shadowWithColor:(NSColor *)color blurRadius:(CGFloat)radius offset:(NSSize)offset;
 @end
